backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2  # Fixed: FastAPI 0.104.1 needs Pydantic 2.x
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# Async Support
aiofiles==23.2.1
# asyncio is built-in, removed from requirements

# File Processing
python-magic==0.4.27
# python-magic-bin==0.4.14  # Only needed on Windows, see instructions below
chardet==5.2.0
filetype==1.2.0

# Data Processing
pandas==2.0.3  # Fixed: 2.1.3 requires Python 3.11+
numpy==1.24.4  # Fixed: Compatible with Python 3.10
python-dateutil==2.8.2
scipy==1.10.1  # Added for analysis

# Machine Learning
scikit-learn==1.3.2
# numba==0.57.1  # Fixed: 0.58.1 requires Python 3.11+

# Parsers - Logs
pyparsing==3.1.1
parse==1.19.1
loguru==0.7.2

# Parsers - Network
scapy==2.5.0
# pyshark==0.6  # Requires Wireshark, optional
dpkt==1.9.8
 # Added for PCAP parsing

# Parsers - Archives
py7zr==0.20.6
rarfile==4.1
# zipfile38 not needed, use built-in zipfile

# Parsers - Documents
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
beautifulsoup4==4.12.2
lxml==4.9.3

# Parsers - Email
mail-parser==3.15.0
extract-msg==0.45.0

# Parsers - Database
sqlite-utils==3.35.2
sqlparse==0.4.4

# Parsers - Forensics (Optional - need special installation)
# pytsk3==20230125  # See special instructions
# pyewf  # See special instructions
pefile==2023.2.7  # PE file analysis
python-registry==1.3.1  # Windows registry

# Analysis Engines
yara-python==4.3.1
# pySigma==0.10.5  # Not available, use sigma-cli instead
stix2==3.0.1
taxii2-client==2.3.0
pymisp==2.4.183  # Added for MISP integration

# External APIs
openai==1.3.7
requests==2.31.0
aiohttp==3.9.0
httpx==0.25.2  # Added for async HTTP

# WebSocket Support
websockets==12.0
python-socketio==5.10.0

# Security & Hashing
cryptography==41.0.7
# hashlib is built-in, removed
pytz==2023.3
pycryptodome==3.19.0  # Added for crypto operations

# Storage
redis==5.0.1
# aiocache==0.12.2  # Deprecated, use redis directly
diskcache==5.6.3  # Added for local caching

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1
tabulate==0.9.0

# JSON handling
orjson==3.9.10
ujson==5.8.0

# Performance
# cython==3.0.6  # Optional, for building extensions

# Platform specific (uncomment as needed)
# Windows Event Log
# python-evtx==0.7.4  # Windows only

# Linux monitoring
# pyinotify==0.9.6  # Linux only

# Additional useful packages
psutil==5.9.6  # System monitoring
watchdog==3.0.0  # File system events
python-crontab==3.0.0  # Cron job parsing
validators==0.22.0  # URL/email validation
arrow==1.3.0  # Better datetime
pyyaml==6.0.1  # YAML support
toml==0.10.2  # TOML support